refactor: AuthServiceIntegrationTest 통합
