+++ backend/src/test/java/com/festago/auth/application/integration/AuthServiceIntegrationTest.java
@@ -1,6 +1,7 @@
 package com.festago.auth.application.integration;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import com.festago.application.integration.ApplicationIntegrationTest;
 import com.festago.auth.application.AuthService;
@@ -8,6 +9,9 @@ import com.festago.auth.domain.SocialType;
 import com.festago.auth.dto.LoginRequest;
 import com.festago.domain.Member;
 import com.festago.domain.MemberRepository;
+import com.festago.support.MemberFixture;
+import jakarta.persistence.EntityManager;
+import java.util.List;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
@@ -23,7 +27,9 @@ class AuthServiceIntegrationTest extends ApplicationIntegrationTest {
     @Autowired
     AuthService authService;
 
-    // TODO #267 이슈 Merge시 MemberService로 이동
+    @Autowired
+    EntityManager entityManager;
+
     @Test
     void 회원이_탈퇴하고_재가입하면_새로운_계정으로_가입() {
         // given
@@ -37,5 +43,41 @@ class AuthServiceIntegrationTest extends ApplicationIntegrationTest {
 
         // then
         assertThat(memberRepository.count()).isEqualTo(1);
+    }
+
+    @Test
+    void 회원탈퇴() {
+        // given
+        Member member = memberRepository.save(MemberFixture.member().build());
+
+        // when
+        authService.deleteMember(member.getId());
+
+        // then
+        String sql = "SELECT * FROM member WHERE id = :memberId AND deleted_at IS NOT NULL";
+        List<Member> deletedMembers = entityManager.createNativeQuery(sql, Member.class)
+            .setParameter("memberId", member.getId())
+            .getResultList();
+
+        assertSoftly(softly -> {
+            softly.assertThat(deletedMembers.size()).isOne();
+            Member actual = deletedMembers.get(0);
+            softly.assertThat(actual.getId()).isEqualTo(member.getId());
+            softly.assertThat(actual.getNickname()).isEqualTo("탈퇴한 회원");
+            softly.assertThat(actual.getProfileImage()).isBlank();
+        });
+    }
+
+
+    @Test
+    void 탈퇴한_회원은_조회되지않는다() {
+        // given
+        Member member = memberRepository.save(MemberFixture.member().build());
+
+        // when
+        authService.deleteMember(member.getId());
+
+        // then
+        assertThat(memberRepository.findById(member.getId())).isEmpty();
     }
 }
